Refactored architecture to be domain driven.
