+++ src/App.js
@@ -3,9 +3,9 @@ import { BrowserRouter } from "react-router-dom";
 import GlobalStyles from "./GlobalStyles";
 import Routes from "./Routes";
 
-import Container from "./components/Container/Container";
-import Header from "./components/Header/Header";
-import SidePanel from "./components/SidePanel/SidePanel";
+import Container from "./components/shared/Container/Container";
+import Header from "./components/shared/Header/Header";
+import SidePanel from "./components/shared/SidePanel/SidePanel";
 import JobProvider from "./context/jobContext";
 
 function App() {
