+++ src/Routes.js
@@ -1,6 +1,6 @@
 import { Routes as RoutesList, Route } from "react-router-dom";
-import JobView from "./components/JobView/JobView";
-import Table from "./components/Table/Table";
+import JobView from "./components/job/JobView/JobView";
+import Table from "./components/jobs/Table/Table";
 
 const Routes = () => {
   return (
